% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{g}
\alias{g}
\title{Helper function to got log ratio of cumulative incidences}
\usage{
g(F)
}
\arguments{
\item{F}{A vector of cumulative incidence estimates with alternating
treatment = 0 and treatment = 1 for each type.}
}
\description{
Helper function to got log ratio of cumulative incidences
}
