% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.ctimeForm}
\alias{get.ctimeForm}
\title{Make a formula that will implement an empirical hazard
estimator when called in estimateCensoring within \code{survtmle}.
See \code{?survtmle} for more details}
\usage{
get.ctimeForm(trt, site, ftime, ftype)
}
\arguments{
\item{trt}{A vector of vaccine assignments}

\item{site}{A vector of geographical site assignments}

\item{ftime}{A vector of failure times}

\item{ftype}{A vector of failure types}
}
\value{
A character formula to be input to \code{glm.ctime} in 
calls to \code{survtmle}
}
\description{
Make a formula that will implement an empirical hazard
estimator when called in estimateCensoring within \code{survtmle}.
See \code{?survtmle} for more details
}
