% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMO}
\alias{getMO}
\title{Get multiply outputed cumulative incidence estimates and 
influence function estimates.}
\usage{
getMO(rslt)
}
\arguments{
\item{rslt}{A list of \code{survtmle} objects}
}
\value{
A list with named entries \code{est} containing averaged
cumulative incidence estimates and \code{ic} containing averaged
influence function estimates.
}
\description{
Get multiply outputed cumulative incidence estimates and 
influence function estimates.
}
