% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grad_g}
\alias{grad_g}
\title{Helper function to compute the gradient for the log ratio
transformation of the cumulative incidence estimates.}
\usage{
grad_g(F)
}
\arguments{
\item{F}{A vector of cumulative incidence estimates with alternating
treatment = 0 and treatment = 1 for each type.}
}
\description{
Helper function to compute the gradient for the log ratio
transformation of the cumulative incidence estimates.
}
