% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeData}
\alias{makeData}
\title{makeData}
\usage{
makeData(n, t0 = 6, ...)
}
\arguments{
\item{n}{Sample size}

\item{t0}{The final timepoint}

\item{...}{Other args, not currently used}
}
\description{
Make the data used in the simulation study
}
