% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeIllustrationPlot}
\alias{makeIllustrationPlot}
\title{Make a plot illustrating the true trend in vaccine efficacy}
\usage{
makeIllustrationPlot(file = "~/Dropbox/Emory/hammingSieve/sim1.pdf")
}
\arguments{
\item{file}{Name of file to save. If \code{NULL}}
}
\description{
Make a plot illustrating the true trend in vaccine efficacy
}
