% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.trend_test}
\alias{print.trend_test}
\title{Print the output of trend_test.}
\usage{
\method{print}{trend_test}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"trend_test"}.}

\item{digits}{Number of digits to round output.}

\item{...}{Other options (not currently used)}
}
\description{
Print the output of trend_test.
}
