% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trend_test}
\alias{trend_test}
\title{Test for a trend in sieve effect across different levels 
of an ordinal failure type.}
\usage{
trend_test(object, level = 0.95)
}
\arguments{
\item{object}{An object of class \code{survtmle}.}

\item{level}{The nominal coverage probability of the confidence interval.}
}
\value{
An object of class \code{"trend_test"}.
\describe{
 \item{\code{alpha}}{The intercept from the projection onto the working model.}
	\item{\code{beta}}{The slope from the projection onto the working model, i.e., 
		  the "trend" parameter.}
	\item{\code{R}}{A vector of the log ratio of cumulative incidences that were 
	  projected onto the working model.}
	\item{\code{se_beta}}{The influence function-based estimate of the standard 
		  error of \code{beta_n}.}
	\item{\code{ci}}{The confidence interval at the requested level for \code{beta_n}.}
	\item{\code{pval}}{The two-sided p-value from the test of the null hypothesis that
		  \code{beta_n} = 0.}
	\item{\code{level}}{The requested level of the confidence interval.}
}
}
\description{
Test for a trend in sieve effect across different levels 
of an ordinal failure type.
}
